Mv CollectionInfo to models.models
