+++ ansible_galaxy/collection_info.py
@@ -4,8 +4,7 @@ import pprint
 
 import yaml
 
-from ansible_galaxy.models.collection_info import \
-    CollectionInfo
+from ansible_galaxy.models.models import CollectionInfo
 
 log = logging.getLogger(__name__)
 pf = pprint.pformat
