+++ ansible_galaxy/models/models.py
@@ -21,3 +21,14 @@ class CollectionArtifactFile(object):
     chksum_sha256 = attr.ib(default=None)
     # name = attr.ib(default=None)
     format_version = attr.ib(default=0.0)
+
+
+# see https://github.com/ansible/galaxy/issues/957
+@attr.s(frozen=True)
+class CollectionInfo(object):
+    namespace = attr.ib()
+    name = attr.ib()
+    version = attr.ib()
+    format_version = attr.ib(default=0.0)
+    author = attr.ib(default=None)
+    license = attr.ib(default=None)
